+++ pinns4bvp/__init__.py
@@ -58,4 +58,4 @@ __all__ = [
     "continue_parameter",
     "solve",
 ]
-__version__ = "0.8.0.dev1"
+__version__ = "0.8.0.dev2"